Fix build issue, and moved test to tests folder
