+++ src/tsconfig.json
@@ -1,0 +1,8 @@
+{
+    "extends": "../tsconfig.json",
+    "compilerOptions": {
+        "rootDir": "./",
+        "outDir": "../dist"
+    },
+    "includes": ["./"]
+}
