+++ tsconfig.json
@@ -8,11 +8,9 @@
         "esModuleInterop": true,
         "target": "ES2020",
         "sourceMap": true,
-        "rootDir": "./",
-        "outDir": "./dist",
-        "baseUrl": "./",
         "incremental": true,
-        "moduleResolution": "node"
+        "moduleResolution": "node",
+        "baseUrl": "./"
     },
-    "exclude": ["src/tests", "jest.config.ts"]
+    "exclude": ["node_modules", "dist"]
 }
